New DSL etc

+++ spec/integration/failure_diagnostics_spec.rb
@@ -17,16 +17,17 @@ RSpec.describe 'Failure diagnostics' do
 		files = {
 			'_layouts' => {
 				'default.html' => '{% if %}'
-			}
+			},
+			'index.md' => "---\nlayout: default\n---\nBroken\n"
 		}
 
 		expect do
 			suppress_expected_build_stderr do
-				build_jekyll_site(files: files) { |_site, _paths| }
+				jekyll_build(files: files) { |_site, _files| }
 			end
 		end.to raise_error(JekyllTestHarness::SiteBuildError) { |error|
 			expect(error.message).to include('Jekyll site build failed:')
-			expect(error.message).to include('keep_site_on_failure: true')
+			expect(error.message).to include('failures: :keep')
 			expect(error.source_path).not_to be_nil
 			expect(error.destination_path).not_to be_nil
 			expect(error.config_snapshot).to be_a(Hash)
@@ -34,24 +35,29 @@ RSpec.describe 'Failure diagnostics' do
 		}
 	end
 
-	it 'retains temporary directories on failure when keep_site_on_failure is enabled' do
+	it 'retains temporary directories on failure when failures mode is keep' do
 		files = {
 			'_layouts' => {
 				'default.html' => '{% if %}'
-			}
+			},
+			'index.md' => "---\nlayout: default\n---\nBroken\n"
 		}
 		temporary_site_root = nil
+		begin
+			JekyllTestHarness.install!(framework: :rspec, failures: :keep)
+			expect do
+				suppress_expected_build_stderr do
+					jekyll_build(files: files) { |_site, _files| }
+				end
+			end.to raise_error(JekyllTestHarness::SiteBuildError) { |error|
+				temporary_site_root = File.dirname(error.source_path)
+			}
 
-		expect do
-			suppress_expected_build_stderr do
-				build_jekyll_site(files: files, keep_site_on_failure: true) { |_site, _paths| }
-			end
-		end.to raise_error(JekyllTestHarness::SiteBuildError) { |error|
-			temporary_site_root = File.dirname(error.source_path)
-		}
-
-		expect(Dir.exist?(temporary_site_root)).to be(true)
-		FileUtils.remove_entry(temporary_site_root)
+			expect(Dir.exist?(temporary_site_root)).to be(true)
+		ensure
+			FileUtils.remove_entry(temporary_site_root) if temporary_site_root && Dir.exist?(temporary_site_root)
+			JekyllTestHarness.install!(framework: :rspec, failures: :clean)
+		end
 	end
 end
